rm unused cert generation script
This was originally landed in the `tokio-tls` repo to support the
`tokio-native-tls` crate's smoke tests. It was never used by the Rustls
code in that repo, but was carried over anyway when we extracted that
code into this repo.

Let's remove it. We can come up with a better solution for the vendored
test certificates we are using.